apps/backend/requirements.txt
flask==3.1.0
pyngrok==7.5.0
torch==2.7.1
torchvision==0.22.1
torchaudio==2.7.1
segmentation-models-pytorch==0.5.0
efficientnet-pytorch==0.7.1
pillow==11.2.1
matplotlib==3.10.1
flask-cors==6.0.1
python-dotenv==1.0.0
